Refactor credential store implementation
- Removed old credential_store.g.dart and credential_store_test.dart files.
- Introduced new credential_config.dart and credential_store.dart files with updated structure.
- Implemented Freezed for data classes AuthProfile and CredentialConfig for better immutability and pattern matching.
- Updated JSON serialization/deserialization methods for the new data structure.
- Enhanced CredentialStore class to manage profiles and configurations more effectively.
- Added comprehensive tests for the new CredentialStore functionality, ensuring proper profile management and configuration persistence.
