+++ apps/genuineci_cli/lib/genuineci_cli.dart
@@ -1,10 +1,10 @@
 library;
 
-export 'src/auth/credential_store.dart';
 export 'src/command_runner.dart';
 export 'src/commands/login_command.dart';
 export 'src/commands/use_command.dart';
 export 'src/config/cli_config.dart';
 export 'src/config/cli_config_data.dart';
+export 'src/credential_store/credential_store.dart';
 export 'src/i18n/i18n.dart';
 export 'src/json_file_store/json_file_store.dart';
